Add crate dependencies and create get_image_files function
Image crate added for its image resizing functionality,
structopt crate added for developing the CLI portion of the application.

The get_image_files function works as follows:

It takes a PathBuf parameter, which is a mutable folder path string.
We initially handle errors with map_err, which maps any errors to a custom ImagixError
and collects all paths in a vector, which we then filter by relevant file extension.